pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "kosmos-ai-scientist"
version = "0.2.0"
description = "Fully autonomous AI scientist powered by Claude for hypothesis generation, experimental design, and iterative discovery"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Kosmos Contributors"}
]
keywords = ["ai", "science", "research", "autonomous", "claude", "discovery"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    # LLM Providers
    "anthropic>=0.40.0",  # Claude/Anthropic provider
    "openai>=1.0.0",      # OpenAI provider and OpenAI-compatible APIs

    # Configuration and validation
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",

    # Database
    "sqlalchemy>=2.0.0",
    "alembic>=1.13.0",

    # HTTP client and retry logic
    "httpx>=0.27.0",
    "tenacity>=8.2.0",

    # Scientific computing
    "numpy>=1.24.0",        # Allow 2.x for pydeseq2 compatibility
    "pandas>=2.0.0",
    "scipy>=1.10.0",
    "scikit-learn>=1.3.0",
    "statsmodels>=0.14.0",  # Statistical power analysis (Phase 4)

    # Visualization
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "plotly>=5.14.0",

    # Literature and knowledge (Phase 1)
    "arxiv>=2.1.0",
    "chromadb>=0.4.0",
    "networkx>=3.1",

    # Phase 2: Literature APIs & Processing
    "semanticscholar>=0.8.0",     # Semantic Scholar API
    "biopython>=1.81",            # PubMed E-utilities
    "pymupdf>=1.23.0",            # PDF text extraction
    "sentence-transformers>=2.2.0", # SPECTER embeddings
    "bibtexparser>=1.4.0",        # BibTeX parsing
    "pybtex>=0.24.0",             # Bibliography generation
    "pikepdf>=8.10.0",            # PDF metadata
    "py2neo>=2021.2.3",           # Neo4j Python driver

    # Phase 9: Multi-Domain Support - Biology
    "pykegg>=0.1.0",              # KEGG pathway database API
    "mygene>=3.2.0",              # Gene annotation queries
    "pyensembl>=2.3.0",           # Ensembl genome database API

    # Phase 9: Multi-Domain Support - Neuroscience
    "pydeseq2>=0.4.0",            # Differential expression analysis

    # Phase 9: Multi-Domain Support - Materials/Physics
    "pymatgen>=2024.1.0",         # Materials analysis and property prediction
    "ase>=3.22.0",                # Atomic Simulation Environment
    "mp-api>=0.41.0",             # Materials Project API client
    "xgboost>=2.0.0",             # Gradient boosting models

    # CLI and UI
    "rich>=13.0.0",
    "click>=8.1.0",
    "typer>=0.9.0",
]

[project.optional-dependencies]
dev = [
    # Testing
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "pytest-timeout>=2.2.0",
    "pytest-xdist>=3.5.0",  # Parallel test execution
    "responses>=0.24.0",     # Mock HTTP responses
    "freezegun>=1.4.0",      # Mock datetime
    "faker>=22.0.0",         # Generate fake data for tests

    # Code quality
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
    "pre-commit>=3.6.0",
]

router = [
    # Optional: Use anthropic_router for CLI routing
    # Users can install this to enable Claude Code CLI support
    "anthropic-router @ git+https://github.com/jimmc414/claude_n_codex_api_proxy.git",
]

domain-extended = [
    # Extended neuroscience capabilities
    "nilearn>=0.10.0",       # Neuroimaging analysis and visualization

    # Physics simulation engines
    "pybullet>=3.2.0",       # Physics engine for simulations
    "simpy>=4.1.0",          # Discrete event simulation

    # ML interpretability (optional due to NumPy 1.x dependency)
    "shap>=0.44.0",          # SHAP analysis (install separately if needed)

    # Chemistry tools (future expansion)
    # "rdkit>=2023.9.0",     # Cheminformatics (requires conda)
    # "openbabel>=3.1.0",    # Chemical file formats (requires system libs)
]

[project.scripts]
kosmos = "kosmos.cli.main:cli_entrypoint"

[project.urls]
Homepage = "https://github.com/jimmc414/Kosmos"
Documentation = "https://github.com/jimmc414/Kosmos/tree/master/docs"
Repository = "https://github.com/jimmc414/Kosmos"
Issues = "https://github.com/jimmc414/Kosmos/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["kosmos*"]
exclude = ["tests*", "docs*", "kosmos-figures*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=kosmos",
    "--cov-report=term-missing",
    "--cov-report=html",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "e2e: marks tests as end-to-end tests",
]

[tool.coverage.run]
source = ["kosmos"]
omit = ["*/tests/*", "*/test_*.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]

[tool.black]
line-length = 100
target-version = ["py311", "py312"]
include = '\.pyi?$'
extend-exclude = '''
/(
  # Directories
  \.git
  | \.mypy_cache
  | \.pytest_cache
  | \.venv
  | venv
  | build
  | dist
  | kosmos-figures
)/
'''

[tool.ruff]
line-length = 100
target-version = "py311"
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex (we'll handle this case-by-case)
]
exclude = [
    ".git",
    "__pycache__",
    ".venv",
    "venv",
    "build",
    "dist",
    "kosmos-figures",
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]  # Allow unused imports in __init__.py

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false  # Start permissive, can tighten later
ignore_missing_imports = true
exclude = [
    'kosmos-figures',
    'tests',
    'docs',
]
